requirements.txt
pandas>=1.3.0
numpy>=1.20.0
matplotlib>=3.4.0
wordcloud>=1.8.0
scikit-learn>=0.24.0
gensim>=4.0.0
nltk>=3.6.0
sentence-transformers>=2.0.0
bertopic>=0.10.0
umap-learn>=0.5.0
hdbscan>=0.8.0
transformers>=4.5.0
torch>=1.9.0 